refactor(frontend): move auth thunks from common/ to web/auth/authThunks
Move loadCurrentUser(), pushLogin(), and pushLogout() out of
currentUserSlice.ts (common/) into a new web/auth/authThunks.ts file,
enforcing the correct dependency direction: common/ must not depend on
web/.

Previously these thunks used a synchronous require() inside closures to
lazily load authClient, but webpack statically resolves synchronous
require() calls, meaning better-auth and authClient were still included
in the desktop bundle. Moving them to web/ ensures the desktop webpack
entrypoint never reaches them.

- Add src/frontend/web/auth/authThunks.ts with the three auth thunks
- Add src/frontend/web/auth/authThunks.test.ts co-located with the thunks
- Remove the thunks and authClient reference from currentUserSlice.ts
- Update imports in MainRouter.tsx, AdminHome.tsx, PublicHome.tsx
- Update jest.config.js moduleNameMapper to intercept ./authClient
  (relative sibling import used by authThunks.ts) as well as the
  existing cross-layer path pattern
