+++ src/frontend/common/state/currentUserSlice.ts
@@ -1,5 +1,4 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { AppDispatch } from "./appState";
 import { User } from "../../../core/models/User";
 import { Locale } from "../../../core/i18n/I18n";
 
@@ -40,48 +39,3 @@ const currentUserSlice = createSlice({
 });
 
 export default currentUserSlice;
-
-// eslint-disable-next-line @typescript-eslint/no-require-imports
-const getAuthClient = () => require("../../web/auth/authClient").authClient as typeof import("../../web/auth/authClient").authClient;
-
-export function loadCurrentUser() {
-  return async (dispatch: AppDispatch) => {
-    const authClient = getAuthClient();
-    const result = await authClient.getSession();
-    if (result?.data?.user) {
-      const { id, admin } = result.data.user as { id: string; admin?: boolean };
-      dispatch(currentUserSlice.actions.setUser({ id, admin: Boolean(admin) }));
-    } else {
-      dispatch(currentUserSlice.actions.setUser(null));
-    }
-  };
-}
-
-export function pushLogin(login: { email: string; password: string }) {
-  return async (dispatch: AppDispatch) => {
-    const authClient = getAuthClient();
-    const result = await authClient.signIn.email({
-      email: login.email,
-      password: login.password,
-      callbackURL: "/",
-    });
-    if (result.error && result.error.status === 401) {
-      dispatch(
-        currentUserSlice.actions.setError(
-          result.error.message ?? "Invalid credentials"
-        )
-      );
-    } else if (result.data?.user) {
-      const { id, admin } = result.data.user as { id: string; admin?: boolean };
-      dispatch(currentUserSlice.actions.setUser({ id, admin: Boolean(admin) }));
-    }
-  };
-}
-
-export function pushLogout() {
-  return async (dispatch: AppDispatch) => {
-    const authClient = getAuthClient();
-    await authClient.signOut();
-    dispatch(currentUserSlice.actions.setUser(null));
-  };
-}
